fix(auth): add validation on authority

+++ backend/auth-core/src/main/java/id/my/rascal/auth/internal/controller/AuthorityController.java
@@ -6,6 +6,8 @@ import id.my.rascal.auth.internal.model.request.AuthorityRequest;
 import id.my.rascal.auth.internal.model.response.AuthorityResponse;
 import id.my.rascal.auth.internal.service.AuthorityService;
 import id.my.rascal.common.ApiResponse;
+import id.my.rascal.common.exception.BadRequestException;
+import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -23,7 +25,7 @@ public class AuthorityController {
     }
 
     @PostMapping
-    public ResponseEntity<?> create(@RequestBody AuthorityRequest request) {
+    public ResponseEntity<?> create(@RequestBody @Valid AuthorityRequest request) {
         AuthorityResponse data = authorityService.create(request);
         return ApiResponse.success(HttpStatus.CREATED, "Authority created", data);
     }
@@ -35,8 +37,13 @@ public class AuthorityController {
     }
 
     @GetMapping
-    public ResponseEntity<?> getAll(Pageable pageable) {
-        Page<AuthorityResponse> page = authorityService.getAllPaged(pageable);
+    public ResponseEntity<?> getAll(
+        @RequestParam(required = false) String name,
+        Pageable pageable
+    ) {
+        Page<AuthorityResponse> page = authorityService
+            .searchActiveAuthorities(name, pageable);
+
         return ApiResponse.paged(
             HttpStatus.OK, "Authorities retrieved",
             page.getContent(),
@@ -49,20 +56,33 @@ public class AuthorityController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> updatePut(@PathVariable Long id, @RequestBody AuthorityPutRequest request) {
+    public ResponseEntity<?> updatePut(
+        @PathVariable Long id,
+        @RequestBody @Valid AuthorityPutRequest request
+    ) {
         AuthorityResponse data = authorityService.update(id, request);
         return ApiResponse.success(HttpStatus.OK, "Authority updated", data);
     }
 
     @PatchMapping("/{id}")
-    public ResponseEntity<?> updatePatch(@PathVariable Long id, @RequestBody AuthorityPatchRequest request) {
+    public ResponseEntity<?> updatePatch(
+        @PathVariable Long id,
+        @RequestBody AuthorityPatchRequest request
+    ) {
+        if (request.isEmptyPatch())
+            throw new BadRequestException("PATCH can't be empty");
+
         AuthorityResponse data = authorityService.update(id, request);
         return ApiResponse.success(HttpStatus.OK, "Authority updated", data);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> delete(@PathVariable Long id) {
+        if (id <= 0)
+            throw new BadRequestException("Invalid ID: " + id);
+
         authorityService.delete(id);
         return ApiResponse.success(HttpStatus.OK, "Authority deleted", null);
     }
+
 }
